feat: add heatmap component
